Updated Routing and refactored login routes
---
- `App.tsx`: Updated to be more modernized and use ProtectedRoutes
- `Header.tsx`: Created file to contain app header
- `Footer.tsx`: Moved this to components folder
- `ProtectedRoute.tsx`: Created a component that would help protect enforce authentication on routes. This would also reduce our repetition and make things cleaner.
- `NavBar.tsx`: Updated it so that it uses constants in `constants/routes.ts instead of hard coded values.
- `routes.ts`: Created file to contain the routes that react-router-dom uses to route our app. This is held in a new folder called `constants`. It's in the top level of the app, so I guess constants here can be used across multiple sections of the app.
- `AppLayout.tsx`: Created layout that organizes top level components like header, navbar, and footer, and then the main content of each page being in the middle.
- `AuthLayout.tsx`: Created a layout to contain user login, registration, and forgot password forms.
- `ForgotPasswordForm`: Created a form when someone forgets their password.
- `Login.tsx`: Deleted the old Login component because it's not needed anymore.
- `LoginForm.tsx` and `RegistrationForm.tsx`: Created separate form components for handling user login and registration. Now the logic is done separately rather than in one `Login.tsx` component.
- `TermsModal.tsx`: Extracted into its own component to make our code less verbose.
- `routes` and `userContext.tsx`: This directory and the file was never used nor referenced, so they were deleted.

---
Our routes are a little more modernized now. We're using layouts, which extracts a lot of the html out of our app and into components. For example, our AppLayout sets up the layout for the entire application. There you see the header, navbar, and footer, and also the main-content where react router will render the pages of our SPA.

As well as this, we improved the maintainability of our routing system. Before our routes were defined in our app, and if you wanted to link to a route, you had to type the exact path, which is prone to error. For example, Login route is "/Login" and in our Navbar we had to write out "/Login" for another link that redirected to the login page. If you mistyped this, then it wouldn't redirect, and imagine you wanted to change "/Login" to "/login" or you changed the path structure. It would be messy. To fix this, we defined the routes in a constant dictionary, so now you can just import the dictionary and reference the route you want, no mistakes possible.

Finally, to make your code reusable we created a ProtectedRoute component in attempts to reduce repetition.

---
We refactored the login section to be 3 different routes for logging in, registering, and also sending a forgot password email.